--- src/server/HelloCodeaholics.Common/Domain/Entities/Location.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HelloCodeaholics.Common.Domain.Entities
-{
-    public class Location
-    {
-        public string Address { get; set; } = string.Empty;
-        public string City { get; set; } = string.Empty;
-        public string State { get; set; } = string.Empty;
-        public int Zip { get; set; }
-    }
-}
