--- src/server/HelloCodeaholics.Common/Domain/Entities/Warehouse.cs
@@ -1,7 +1,0 @@
-﻿namespace HelloCodeaholics.Common.Domain.Entities;
-
-public class Warehouse : Location
-{
-    public int Id { get; set; }
-    public string Name { get; set; } = string.Empty;
-}
