+++ src/server/HelloCodeaholics.Data/Entities/Pharmacist.cs
@@ -4,10 +4,11 @@ public class Pharmacist
 {
     public int Id { get; set; }
     public int PharmacyId { get; set; }
-    public string Name { get; set; } = null!;
-    public string PrimaryDrug { get; set; } = null!;
+    public string FirstName { get; set; } = null!;
+    public string LastName { get; set; } = null!;
     public int Age { get; set; }
-    public DateTime HireDate { get; set; }
+    public string PrimaryDrugSold { get; set; } = null!;
+    public DateTime DateOfHire { get; set; }
     public virtual Pharmacy Pharmacy { get; set; } = null!;
 
 }
